Now taking effective timeout taking as minimum of global timeout and session timeout

+++ include/Base_Session.h
@@ -82,7 +82,7 @@ class Base_Session {
 	int pending_connect;
 	enum proxysql_session_type session_type;
 	int warning_in_hg;
-	int wait_timeout;   //in milliseconds
+	unsigned long long wait_timeout;   //in milliseconds
 
 	// bool
 	bool autocommit;
